refactor(memory): split runtime backend query ranking domains (#1686)
