+++ crates/tau-memory/src/runtime.rs
@@ -1,14 +1,27 @@
-use std::collections::{BTreeMap, BTreeSet, HashMap};
-use std::fs::{self, OpenOptions};
-use std::io::{BufRead, BufReader, Write};
 use std::path::{Path, PathBuf};
-use std::time::{Instant, SystemTime, UNIX_EPOCH};
+use std::time::{SystemTime, UNIX_EPOCH};
 
 use anyhow::{anyhow, bail, Context, Result};
-use rusqlite::{params, Connection};
+use rusqlite::params;
 use serde::{Deserialize, Serialize};
 
 use crate::memory_contract::{MemoryEntry, MemoryScope};
+
+mod backend;
+mod query;
+mod ranking;
+
+use backend::{
+    append_record_jsonl, append_record_sqlite, initialize_memory_sqlite_schema, load_records_jsonl,
+    load_records_sqlite, open_memory_sqlite_connection, resolve_memory_backend,
+};
+pub use ranking::{
+    cosine_similarity, embed_text_vector, rank_text_candidates, rank_text_candidates_bm25,
+};
+use ranking::{
+    reciprocal_rank_fuse, record_search_text, record_search_text_for_entry,
+    resize_and_normalize_embedding,
+};
 
 const MEMORY_RUNTIME_SCHEMA_VERSION: u32 = 1;
 const MEMORY_RUNTIME_ENTRIES_FILE_NAME: &str = "entries.jsonl";
@@ -408,438 +421,6 @@ impl FileMemoryStore {
         Ok(MemoryWriteResult { record, created })
     }
 
-    /// Reads the latest record for `memory_id`, optionally constrained by `scope_filter`.
-    pub fn read_entry(
-        &self,
-        memory_id: &str,
-        scope_filter: Option<&MemoryScopeFilter>,
-    ) -> Result<Option<RuntimeMemoryRecord>> {
-        let normalized_memory_id = memory_id.trim();
-        if normalized_memory_id.is_empty() {
-            bail!("memory_id must not be empty");
-        }
-        let records = self.load_latest_records()?;
-        Ok(records.into_iter().find(|record| {
-            record.entry.memory_id == normalized_memory_id
-                && scope_filter
-                    .map(|filter| filter.matches_scope(&record.scope))
-                    .unwrap_or(true)
-        }))
-    }
-
-    /// Returns latest records filtered by scope and bounded by `limit`.
-    pub fn list_latest_records(
-        &self,
-        scope_filter: Option<&MemoryScopeFilter>,
-        limit: usize,
-    ) -> Result<Vec<RuntimeMemoryRecord>> {
-        if limit == 0 {
-            return Ok(Vec::new());
-        }
-        let mut records = self.load_latest_records()?;
-        if let Some(filter) = scope_filter {
-            records.retain(|record| filter.matches_scope(&record.scope));
-        }
-        records.truncate(limit);
-        Ok(records)
-    }
-
-    /// Performs deterministic semantic search over latest records.
-    pub fn search(&self, query: &str, options: &MemorySearchOptions) -> Result<MemorySearchResult> {
-        let normalized_query = query.trim();
-        if normalized_query.is_empty() {
-            bail!("query must not be empty");
-        }
-
-        let mut migrated_entries = 0usize;
-        let mut embedding_reason_code = MEMORY_EMBEDDING_REASON_HASH_ONLY.to_string();
-        if options.enable_embedding_migration {
-            let current = self.list_latest_records(Some(&options.scope), usize::MAX)?;
-            match self.migrate_records_to_provider_embeddings(&current) {
-                Ok(migrated) => {
-                    migrated_entries = migrated;
-                }
-                Err(_) => {
-                    embedding_reason_code = MEMORY_EMBEDDING_REASON_PROVIDER_FAILED.to_string();
-                }
-            }
-        }
-
-        let records = self.list_latest_records(Some(&options.scope), usize::MAX)?;
-        let by_memory_id = records
-            .into_iter()
-            .map(|record| (record.entry.memory_id.clone(), record))
-            .collect::<HashMap<_, _>>();
-
-        let query_embedding_started = Instant::now();
-        let computed_query =
-            self.compute_embedding(normalized_query, options.embedding_dimensions, true);
-        let query_embedding_latency_ms = query_embedding_started.elapsed().as_millis() as u64;
-        if computed_query.reason_code != MEMORY_EMBEDDING_REASON_HASH_ONLY {
-            embedding_reason_code = computed_query.reason_code.clone();
-        }
-        let has_query_embedding = computed_query
-            .vector
-            .iter()
-            .any(|component| *component != 0.0);
-
-        let ranking_started = Instant::now();
-        let mut vector_ranked = if has_query_embedding {
-            by_memory_id
-                .iter()
-                .filter_map(|(memory_id, record)| {
-                    let candidate_embedding = if record.embedding_vector.is_empty() {
-                        embed_text_vector(
-                            record_search_text(record).as_str(),
-                            options.embedding_dimensions,
-                        )
-                    } else {
-                        resize_and_normalize_embedding(
-                            record.embedding_vector.as_slice(),
-                            options.embedding_dimensions,
-                        )
-                    };
-                    let score = cosine_similarity(&computed_query.vector, &candidate_embedding);
-                    if score >= options.min_similarity {
-                        Some(RankedTextMatch {
-                            key: memory_id.clone(),
-                            text: record_search_text(record),
-                            score,
-                        })
-                    } else {
-                        None
-                    }
-                })
-                .collect::<Vec<_>>()
-        } else {
-            Vec::new()
-        };
-        vector_ranked.sort_by(|left, right| {
-            right
-                .score
-                .total_cmp(&left.score)
-                .then_with(|| left.key.cmp(&right.key))
-        });
-
-        let mut lexical_ranking_latency_ms = 0u64;
-        let mut lexical_ranked = Vec::new();
-        if options.enable_hybrid_retrieval {
-            let lexical_started = Instant::now();
-            lexical_ranked = rank_text_candidates_bm25(
-                normalized_query,
-                by_memory_id
-                    .iter()
-                    .map(|(memory_id, record)| RankedTextCandidate {
-                        key: memory_id.clone(),
-                        text: record_search_text(record),
-                    })
-                    .collect::<Vec<_>>(),
-                by_memory_id.len(),
-                options.bm25_k1,
-                options.bm25_b,
-                options.bm25_min_score,
-            );
-            lexical_ranking_latency_ms = lexical_started.elapsed().as_millis() as u64;
-        }
-
-        let mut vector_scores = HashMap::new();
-        for item in &vector_ranked {
-            vector_scores.insert(item.key.clone(), item.score);
-        }
-        let mut lexical_scores = HashMap::new();
-        for item in &lexical_ranked {
-            lexical_scores.insert(item.key.clone(), item.score);
-        }
-
-        let vector_candidates = vector_ranked.len();
-        let lexical_candidates = lexical_ranked.len();
-
-        let mut fusion_latency_ms = 0u64;
-        let mut fused_scores = HashMap::new();
-        let mut ranked = if options.enable_hybrid_retrieval {
-            let fusion_started = Instant::now();
-            let fused = reciprocal_rank_fuse(
-                &vector_ranked,
-                &lexical_ranked,
-                options.limit,
-                options.rrf_k,
-                options.rrf_vector_weight,
-                options.rrf_lexical_weight,
-            );
-            fusion_latency_ms = fusion_started.elapsed().as_millis() as u64;
-            for item in &fused {
-                fused_scores.insert(item.key.clone(), item.score);
-            }
-            fused
-        } else {
-            vector_ranked.clone()
-        };
-
-        ranked.truncate(options.limit);
-        let ranking_latency_ms = ranking_started.elapsed().as_millis() as u64;
-
-        let mut matches = Vec::with_capacity(ranked.len());
-        for item in &ranked {
-            let Some(record) = by_memory_id.get(&item.key) else {
-                continue;
-            };
-            matches.push(MemorySearchMatch {
-                memory_id: record.entry.memory_id.clone(),
-                score: item.score,
-                vector_score: vector_scores.get(item.key.as_str()).copied(),
-                lexical_score: lexical_scores.get(item.key.as_str()).copied(),
-                fused_score: options
-                    .enable_hybrid_retrieval
-                    .then(|| fused_scores.get(item.key.as_str()).copied())
-                    .flatten(),
-                scope: record.scope.clone(),
-                summary: record.entry.summary.clone(),
-                tags: record.entry.tags.clone(),
-                facts: record.entry.facts.clone(),
-                source_event_key: record.entry.source_event_key.clone(),
-                embedding_source: record.embedding_source.clone(),
-                embedding_model: record.embedding_model.clone(),
-            });
-        }
-        let selected = matches
-            .iter()
-            .map(|item| item.memory_id.as_str())
-            .collect::<BTreeSet<_>>();
-
-        let baseline_hash_overlap = if options.benchmark_against_hash {
-            let baseline = rank_text_candidates(
-                normalized_query,
-                by_memory_id
-                    .iter()
-                    .map(|(memory_id, record)| RankedTextCandidate {
-                        key: memory_id.clone(),
-                        text: record_search_text(record),
-                    })
-                    .collect::<Vec<_>>(),
-                options.limit,
-                options.embedding_dimensions,
-                options.min_similarity,
-            );
-            Some(
-                baseline
-                    .into_iter()
-                    .filter(|candidate| selected.contains(candidate.key.as_str()))
-                    .count(),
-            )
-        } else {
-            None
-        };
-        let baseline_vector_overlap = if options.benchmark_against_vector_only {
-            Some(
-                vector_ranked
-                    .iter()
-                    .take(options.limit)
-                    .filter(|candidate| selected.contains(candidate.key.as_str()))
-                    .count(),
-            )
-        } else {
-            None
-        };
-        let retrieval_backend = if options.enable_hybrid_retrieval {
-            MEMORY_RETRIEVAL_BACKEND_HYBRID_BM25_RRF.to_string()
-        } else {
-            MEMORY_RETRIEVAL_BACKEND_VECTOR_ONLY.to_string()
-        };
-        let retrieval_reason_code = if options.enable_hybrid_retrieval {
-            MEMORY_RETRIEVAL_REASON_HYBRID_ENABLED.to_string()
-        } else {
-            MEMORY_RETRIEVAL_REASON_VECTOR_ONLY.to_string()
-        };
-
-        Ok(MemorySearchResult {
-            query: normalized_query.to_string(),
-            scanned: by_memory_id.len(),
-            returned: matches.len(),
-            retrieval_backend,
-            retrieval_reason_code,
-            embedding_backend: computed_query.backend,
-            embedding_reason_code,
-            migrated_entries,
-            query_embedding_latency_ms,
-            ranking_latency_ms,
-            lexical_ranking_latency_ms,
-            fusion_latency_ms,
-            vector_candidates,
-            lexical_candidates,
-            baseline_hash_overlap,
-            baseline_vector_overlap,
-            matches,
-        })
-    }
-
-    fn compute_embedding(
-        &self,
-        text: &str,
-        dimensions: usize,
-        prefer_provider: bool,
-    ) -> ComputedEmbedding {
-        if prefer_provider {
-            if let Some(config) = &self.embedding_provider {
-                let provider = config.provider.trim().to_ascii_lowercase();
-                if provider == "openai" || provider == "openai-compatible" {
-                    if let Ok(vectors) =
-                        embed_text_vectors_via_provider(&[text.to_string()], dimensions, config)
-                    {
-                        if let Some(first) = vectors.first() {
-                            return ComputedEmbedding {
-                                vector: first.clone(),
-                                backend: MEMORY_EMBEDDING_SOURCE_PROVIDER.to_string(),
-                                model: Some(config.model.clone()),
-                                reason_code: MEMORY_EMBEDDING_REASON_PROVIDER_SUCCESS.to_string(),
-                            };
-                        }
-                    }
-                    return ComputedEmbedding {
-                        vector: embed_text_vector(text, dimensions),
-                        backend: MEMORY_EMBEDDING_SOURCE_HASH.to_string(),
-                        model: None,
-                        reason_code: MEMORY_EMBEDDING_REASON_PROVIDER_FAILED.to_string(),
-                    };
-                }
-            }
-        }
-
-        ComputedEmbedding {
-            vector: embed_text_vector(text, dimensions),
-            backend: MEMORY_EMBEDDING_SOURCE_HASH.to_string(),
-            model: None,
-            reason_code: MEMORY_EMBEDDING_REASON_HASH_ONLY.to_string(),
-        }
-    }
-
-    fn migrate_records_to_provider_embeddings(
-        &self,
-        records: &[RuntimeMemoryRecord],
-    ) -> Result<usize> {
-        let Some(config) = &self.embedding_provider else {
-            return Ok(0);
-        };
-        let provider = config.provider.trim().to_ascii_lowercase();
-        if provider != "openai" && provider != "openai-compatible" {
-            return Ok(0);
-        }
-
-        let to_migrate = records
-            .iter()
-            .filter(|record| {
-                record.embedding_vector.is_empty()
-                    || record
-                        .embedding_source
-                        .starts_with(MEMORY_EMBEDDING_SOURCE_HASH)
-            })
-            .cloned()
-            .collect::<Vec<_>>();
-        if to_migrate.is_empty() {
-            return Ok(0);
-        }
-
-        let inputs = to_migrate
-            .iter()
-            .map(record_search_text)
-            .collect::<Vec<_>>();
-        let vectors = embed_text_vectors_via_provider(&inputs, config.dimensions, config)
-            .map_err(|error| anyhow::anyhow!(error))?;
-
-        let mut migrated = 0usize;
-        for (record, vector) in to_migrate.into_iter().zip(vectors.into_iter()) {
-            let migrated_record = RuntimeMemoryRecord {
-                schema_version: MEMORY_RUNTIME_SCHEMA_VERSION,
-                updated_unix_ms: current_unix_timestamp_ms(),
-                scope: record.scope,
-                entry: record.entry,
-                embedding_source: MEMORY_EMBEDDING_SOURCE_PROVIDER.to_string(),
-                embedding_model: Some(config.model.clone()),
-                embedding_vector: vector,
-                embedding_reason_code: MEMORY_EMBEDDING_REASON_PROVIDER_SUCCESS.to_string(),
-            };
-            self.append_record_backend(&migrated_record)?;
-            migrated = migrated.saturating_add(1);
-        }
-
-        Ok(migrated)
-    }
-
-    /// Returns a hierarchical workspace/channel/actor tree for latest records.
-    pub fn tree(&self) -> Result<MemoryTree> {
-        let records = self.load_latest_records()?;
-        let mut by_scope = BTreeMap::<String, BTreeMap<String, BTreeMap<String, usize>>>::new();
-
-        for record in records {
-            let workspace = record.scope.workspace_id;
-            let channel = record.scope.channel_id;
-            let actor = record.scope.actor_id;
-            *by_scope
-                .entry(workspace)
-                .or_default()
-                .entry(channel)
-                .or_default()
-                .entry(actor)
-                .or_default() += 1;
-        }
-
-        let mut total_entries = 0usize;
-        let mut workspaces = Vec::with_capacity(by_scope.len());
-        for (workspace_id, channels) in by_scope {
-            let mut workspace_count = 0usize;
-            let mut channel_nodes = Vec::with_capacity(channels.len());
-            for (channel_id, actors) in channels {
-                let mut channel_count = 0usize;
-                let mut actor_nodes = Vec::with_capacity(actors.len());
-                for (actor_id, actor_count) in actors {
-                    channel_count = channel_count.saturating_add(actor_count);
-                    actor_nodes.push(MemoryTreeNode {
-                        id: actor_id,
-                        level: "actor".to_string(),
-                        entry_count: actor_count,
-                        children: Vec::new(),
-                    });
-                }
-                workspace_count = workspace_count.saturating_add(channel_count);
-                channel_nodes.push(MemoryTreeNode {
-                    id: channel_id,
-                    level: "channel".to_string(),
-                    entry_count: channel_count,
-                    children: actor_nodes,
-                });
-            }
-            total_entries = total_entries.saturating_add(workspace_count);
-            workspaces.push(MemoryTreeNode {
-                id: workspace_id,
-                level: "workspace".to_string(),
-                entry_count: workspace_count,
-                children: channel_nodes,
-            });
-        }
-
-        Ok(MemoryTree {
-            total_entries,
-            workspaces,
-        })
-    }
-
-    fn load_latest_records(&self) -> Result<Vec<RuntimeMemoryRecord>> {
-        let records = self.load_records_backend()?;
-        let mut seen = BTreeSet::new();
-        let mut latest = Vec::new();
-        for record in records.into_iter().rev() {
-            if seen.insert(record.entry.memory_id.clone()) {
-                latest.push(record);
-            }
-        }
-        latest.sort_by(|left, right| {
-            right
-                .updated_unix_ms
-                .cmp(&left.updated_unix_ms)
-                .then_with(|| left.entry.memory_id.cmp(&right.entry.memory_id))
-        });
-        Ok(latest)
-    }
-
     fn ensure_backend_ready(&self) -> Result<()> {
         if let Some(error) = &self.backend_init_error {
             bail!(
@@ -950,359 +531,6 @@ impl FileMemoryStore {
     }
 }
 
-/// Ranks text candidates using deterministic hash embeddings and cosine similarity.
-pub fn rank_text_candidates(
-    query: &str,
-    candidates: Vec<RankedTextCandidate>,
-    limit: usize,
-    dimensions: usize,
-    min_similarity: f32,
-) -> Vec<RankedTextMatch> {
-    if limit == 0 {
-        return Vec::new();
-    }
-    let normalized_query = query.trim();
-    if normalized_query.is_empty() {
-        return Vec::new();
-    }
-
-    let query_embedding = embed_text_vector(normalized_query, dimensions);
-    if query_embedding.iter().all(|component| *component == 0.0) {
-        return Vec::new();
-    }
-
-    let mut matches = candidates
-        .into_iter()
-        .filter_map(|candidate| {
-            let candidate_embedding = embed_text_vector(candidate.text.as_str(), dimensions);
-            let score = cosine_similarity(&query_embedding, &candidate_embedding);
-            if score >= min_similarity {
-                Some(RankedTextMatch {
-                    key: candidate.key,
-                    text: candidate.text,
-                    score,
-                })
-            } else {
-                None
-            }
-        })
-        .collect::<Vec<_>>();
-    matches.sort_by(|left, right| {
-        right
-            .score
-            .total_cmp(&left.score)
-            .then_with(|| left.key.cmp(&right.key))
-    });
-    matches.truncate(limit);
-    matches
-}
-
-/// Ranks text candidates using BM25 lexical scoring.
-pub fn rank_text_candidates_bm25(
-    query: &str,
-    candidates: Vec<RankedTextCandidate>,
-    limit: usize,
-    k1: f32,
-    b: f32,
-    min_score: f32,
-) -> Vec<RankedTextMatch> {
-    if limit == 0 || candidates.is_empty() {
-        return Vec::new();
-    }
-
-    let query_tokens = tokenize_text(query);
-    if query_tokens.is_empty() {
-        return Vec::new();
-    }
-
-    let safe_k1 = k1.max(0.1);
-    let safe_b = b.clamp(0.0, 1.0);
-    let safe_min_score = min_score.max(0.0);
-
-    let mut corpus_tokens = Vec::with_capacity(candidates.len());
-    let mut doc_frequencies = HashMap::<String, usize>::new();
-    let mut total_doc_len = 0usize;
-    for candidate in &candidates {
-        let tokens = tokenize_text(candidate.text.as_str());
-        total_doc_len = total_doc_len.saturating_add(tokens.len());
-        let unique_terms = tokens.iter().cloned().collect::<BTreeSet<_>>();
-        for term in unique_terms {
-            *doc_frequencies.entry(term).or_default() += 1;
-        }
-        corpus_tokens.push(tokens);
-    }
-
-    let doc_count = candidates.len() as f32;
-    let average_doc_len = (total_doc_len as f32 / doc_count).max(1.0);
-    let mut matches = Vec::new();
-    for (candidate, tokens) in candidates.into_iter().zip(corpus_tokens.into_iter()) {
-        if tokens.is_empty() {
-            continue;
-        }
-        let mut term_frequencies = HashMap::<String, usize>::new();
-        for token in tokens {
-            *term_frequencies.entry(token).or_default() += 1;
-        }
-
-        let doc_len = term_frequencies.values().sum::<usize>() as f32;
-        let mut score = 0.0f32;
-        for term in &query_tokens {
-            let term_frequency = *term_frequencies.get(term.as_str()).unwrap_or(&0) as f32;
-            if term_frequency <= 0.0 {
-                continue;
-            }
-            let doc_frequency = *doc_frequencies.get(term.as_str()).unwrap_or(&0) as f32;
-            if doc_frequency <= 0.0 {
-                continue;
-            }
-            let idf = (((doc_count - doc_frequency + 0.5) / (doc_frequency + 0.5)) + 1.0).ln();
-            let normalization = safe_k1 * (1.0 - safe_b + safe_b * (doc_len / average_doc_len));
-            let denominator = (term_frequency + normalization).max(f32::EPSILON);
-            score += idf * ((term_frequency * (safe_k1 + 1.0)) / denominator);
-        }
-
-        if score >= safe_min_score {
-            matches.push(RankedTextMatch {
-                key: candidate.key,
-                text: candidate.text,
-                score,
-            });
-        }
-    }
-
-    matches.sort_by(|left, right| {
-        right
-            .score
-            .total_cmp(&left.score)
-            .then_with(|| left.key.cmp(&right.key))
-    });
-    matches.truncate(limit);
-    matches
-}
-
-fn reciprocal_rank_fuse(
-    vector_ranked: &[RankedTextMatch],
-    lexical_ranked: &[RankedTextMatch],
-    limit: usize,
-    rrf_k: usize,
-    vector_weight: f32,
-    lexical_weight: f32,
-) -> Vec<RankedTextMatch> {
-    if limit == 0 {
-        return Vec::new();
-    }
-    let safe_rrf_k = rrf_k.max(1) as f32;
-    let safe_vector_weight = vector_weight.max(0.0);
-    let safe_lexical_weight = lexical_weight.max(0.0);
-
-    let mut scores = HashMap::<String, f32>::new();
-    let mut texts = HashMap::<String, String>::new();
-    for (rank, candidate) in vector_ranked.iter().enumerate() {
-        let contribution = safe_vector_weight / (safe_rrf_k + rank as f32 + 1.0);
-        *scores.entry(candidate.key.clone()).or_default() += contribution;
-        texts
-            .entry(candidate.key.clone())
-            .or_insert_with(|| candidate.text.clone());
-    }
-    for (rank, candidate) in lexical_ranked.iter().enumerate() {
-        let contribution = safe_lexical_weight / (safe_rrf_k + rank as f32 + 1.0);
-        *scores.entry(candidate.key.clone()).or_default() += contribution;
-        texts
-            .entry(candidate.key.clone())
-            .or_insert_with(|| candidate.text.clone());
-    }
-
-    let mut fused = scores
-        .into_iter()
-        .filter_map(|(key, score)| {
-            texts.get(key.as_str()).map(|text| RankedTextMatch {
-                key,
-                text: text.clone(),
-                score,
-            })
-        })
-        .collect::<Vec<_>>();
-    fused.sort_by(|left, right| {
-        right
-            .score
-            .total_cmp(&left.score)
-            .then_with(|| left.key.cmp(&right.key))
-    });
-    fused.truncate(limit);
-    fused
-}
-
-/// Converts text to a normalized fixed-size vector using FNV-1a token hashing.
-pub fn embed_text_vector(text: &str, dimensions: usize) -> Vec<f32> {
-    let dimensions = dimensions.max(1);
-    let mut vector = vec![0.0f32; dimensions];
-    for token in tokenize_text(text) {
-        let hash = fnv1a_hash(token.as_bytes());
-        let index = (hash as usize) % dimensions;
-        let sign = if (hash & 1) == 0 { 1.0 } else { -1.0 };
-        vector[index] += sign;
-    }
-
-    let magnitude = vector
-        .iter()
-        .map(|component| component * component)
-        .sum::<f32>()
-        .sqrt();
-    if magnitude > 0.0 {
-        for component in &mut vector {
-            *component /= magnitude;
-        }
-    }
-    vector
-}
-
-fn embed_text_vectors_via_provider(
-    inputs: &[String],
-    dimensions: usize,
-    config: &MemoryEmbeddingProviderConfig,
-) -> Result<Vec<Vec<f32>>, String> {
-    if inputs.is_empty() {
-        return Ok(Vec::new());
-    }
-
-    let timeout_ms = config.timeout_ms.max(1);
-    let api_base = config.api_base.trim_end_matches('/');
-    if api_base.is_empty() {
-        return Err("embedding api_base must not be empty".to_string());
-    }
-    let client = reqwest::blocking::Client::builder()
-        .timeout(std::time::Duration::from_millis(timeout_ms))
-        .build()
-        .map_err(|error| format!("failed to build embedding client: {error}"))?;
-    let response = client
-        .post(format!("{api_base}/embeddings"))
-        .bearer_auth(config.api_key.as_str())
-        .json(&serde_json::json!({
-            "model": config.model,
-            "input": inputs,
-        }))
-        .send()
-        .map_err(|error| format!("embedding request failed: {error}"))?;
-
-    if !response.status().is_success() {
-        let status = response.status();
-        let body = response.text().unwrap_or_default();
-        return Err(format!(
-            "embedding request failed with status {}: {}",
-            status.as_u16(),
-            body.chars().take(240).collect::<String>()
-        ));
-    }
-
-    let payload = response
-        .json::<serde_json::Value>()
-        .map_err(|error| format!("failed to parse embedding response json: {error}"))?;
-    let data = payload
-        .get("data")
-        .and_then(serde_json::Value::as_array)
-        .ok_or_else(|| "embedding response missing data array".to_string())?;
-    if data.len() != inputs.len() {
-        return Err(format!(
-            "embedding response size mismatch: expected {}, got {}",
-            inputs.len(),
-            data.len()
-        ));
-    }
-
-    let mut vectors = Vec::with_capacity(data.len());
-    for item in data {
-        let raw_embedding = item
-            .get("embedding")
-            .and_then(serde_json::Value::as_array)
-            .ok_or_else(|| "embedding item missing embedding array".to_string())?;
-        let parsed = raw_embedding
-            .iter()
-            .map(|component| {
-                component
-                    .as_f64()
-                    .map(|value| value as f32)
-                    .ok_or_else(|| "embedding component must be numeric".to_string())
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-        vectors.push(resize_and_normalize_embedding(&parsed, dimensions));
-    }
-    Ok(vectors)
-}
-
-fn resize_and_normalize_embedding(values: &[f32], dimensions: usize) -> Vec<f32> {
-    let dimensions = dimensions.max(1);
-    let mut resized = vec![0.0f32; dimensions];
-    for (index, value) in values.iter().enumerate() {
-        let bucket = index % dimensions;
-        resized[bucket] += *value;
-    }
-
-    let magnitude = resized
-        .iter()
-        .map(|component| component * component)
-        .sum::<f32>()
-        .sqrt();
-    if magnitude > 0.0 {
-        for component in &mut resized {
-            *component /= magnitude;
-        }
-    }
-    resized
-}
-
-fn tokenize_text(text: &str) -> Vec<String> {
-    text.split(|character: char| !character.is_alphanumeric())
-        .filter(|token| !token.is_empty())
-        .map(|token| token.to_ascii_lowercase())
-        .collect::<Vec<_>>()
-}
-
-/// Computes cosine similarity for equal-length vectors.
-pub fn cosine_similarity(left: &[f32], right: &[f32]) -> f32 {
-    if left.len() != right.len() {
-        return 0.0;
-    }
-    left.iter()
-        .zip(right)
-        .map(|(left, right)| left * right)
-        .sum()
-}
-
-fn fnv1a_hash(bytes: &[u8]) -> u64 {
-    const FNV_OFFSET_BASIS: u64 = 0xcbf29ce484222325;
-    const FNV_PRIME: u64 = 0x100000001b3;
-    let mut hash = FNV_OFFSET_BASIS;
-    for byte in bytes {
-        hash ^= *byte as u64;
-        hash = hash.wrapping_mul(FNV_PRIME);
-    }
-    hash
-}
-
-fn record_search_text(record: &RuntimeMemoryRecord) -> String {
-    let mut parts = Vec::with_capacity(3);
-    parts.push(record.entry.summary.clone());
-    if !record.entry.tags.is_empty() {
-        parts.push(record.entry.tags.join(" "));
-    }
-    if !record.entry.facts.is_empty() {
-        parts.push(record.entry.facts.join(" "));
-    }
-    parts.join("\n")
-}
-
-fn record_search_text_for_entry(entry: &MemoryEntry) -> String {
-    let mut parts = Vec::with_capacity(3);
-    parts.push(entry.summary.clone());
-    if !entry.tags.is_empty() {
-        parts.push(entry.tags.join(" "));
-    }
-    if !entry.facts.is_empty() {
-        parts.push(entry.facts.join(" "));
-    }
-    parts.join("\n")
-}
-
 fn normalize_scope(scope: &MemoryScope) -> MemoryScope {
     MemoryScope {
         workspace_id: normalize_scope_component(
@@ -1355,265 +583,6 @@ fn normalize_entry(entry: MemoryEntry) -> Result<MemoryEntry> {
         recency_weight_bps: entry.recency_weight_bps,
         confidence_bps: entry.confidence_bps,
     })
-}
-
-fn resolve_memory_backend(root_dir: &Path) -> ResolvedMemoryBackend {
-    let env_backend = std::env::var(MEMORY_BACKEND_ENV)
-        .ok()
-        .map(|value| value.trim().to_ascii_lowercase());
-    let env_backend = env_backend.as_deref().unwrap_or("auto");
-
-    if env_backend != "auto" && env_backend != "jsonl" && env_backend != "sqlite" {
-        let mut inferred = infer_memory_backend(root_dir);
-        inferred.reason_code = MEMORY_STORAGE_REASON_ENV_INVALID_FALLBACK.to_string();
-        return inferred;
-    }
-
-    if env_backend == "jsonl" {
-        return ResolvedMemoryBackend {
-            backend: MemoryStorageBackend::Jsonl,
-            storage_path: Some(resolve_jsonl_path(root_dir)),
-            reason_code: MEMORY_STORAGE_REASON_ENV_JSONL.to_string(),
-        };
-    }
-
-    if env_backend == "sqlite" {
-        return ResolvedMemoryBackend {
-            backend: MemoryStorageBackend::Sqlite,
-            storage_path: Some(resolve_sqlite_path(root_dir)),
-            reason_code: MEMORY_STORAGE_REASON_ENV_SQLITE.to_string(),
-        };
-    }
-
-    let mut inferred = infer_memory_backend(root_dir);
-    if std::env::var(MEMORY_BACKEND_ENV).is_ok() {
-        inferred.reason_code = MEMORY_STORAGE_REASON_ENV_AUTO.to_string();
-    }
-    inferred
-}
-
-fn infer_memory_backend(root_dir: &Path) -> ResolvedMemoryBackend {
-    let extension = root_dir
-        .extension()
-        .and_then(|value| value.to_str())
-        .map(|value| value.to_ascii_lowercase());
-
-    if matches!(extension.as_deref(), Some("jsonl")) {
-        return ResolvedMemoryBackend {
-            backend: MemoryStorageBackend::Jsonl,
-            storage_path: Some(root_dir.to_path_buf()),
-            reason_code: MEMORY_STORAGE_REASON_PATH_JSONL.to_string(),
-        };
-    }
-    if matches!(extension.as_deref(), Some("sqlite" | "db")) {
-        return ResolvedMemoryBackend {
-            backend: MemoryStorageBackend::Sqlite,
-            storage_path: Some(root_dir.to_path_buf()),
-            reason_code: MEMORY_STORAGE_REASON_PATH_SQLITE.to_string(),
-        };
-    }
-
-    if root_dir.exists() && root_dir.is_file() {
-        if looks_like_sqlite_file(root_dir) {
-            return ResolvedMemoryBackend {
-                backend: MemoryStorageBackend::Sqlite,
-                storage_path: Some(root_dir.to_path_buf()),
-                reason_code: MEMORY_STORAGE_REASON_EXISTING_SQLITE.to_string(),
-            };
-        }
-        return ResolvedMemoryBackend {
-            backend: MemoryStorageBackend::Jsonl,
-            storage_path: Some(root_dir.to_path_buf()),
-            reason_code: MEMORY_STORAGE_REASON_EXISTING_JSONL.to_string(),
-        };
-    }
-
-    let sqlite_candidate = root_dir.join(MEMORY_RUNTIME_ENTRIES_SQLITE_FILE_NAME);
-    if sqlite_candidate.exists() {
-        return ResolvedMemoryBackend {
-            backend: MemoryStorageBackend::Sqlite,
-            storage_path: Some(sqlite_candidate),
-            reason_code: MEMORY_STORAGE_REASON_EXISTING_SQLITE.to_string(),
-        };
-    }
-
-    let jsonl_candidate = root_dir.join(MEMORY_RUNTIME_ENTRIES_FILE_NAME);
-    if jsonl_candidate.exists() {
-        return ResolvedMemoryBackend {
-            backend: MemoryStorageBackend::Sqlite,
-            storage_path: Some(resolve_sqlite_path(root_dir)),
-            reason_code: MEMORY_STORAGE_REASON_EXISTING_JSONL.to_string(),
-        };
-    }
-
-    ResolvedMemoryBackend {
-        backend: MemoryStorageBackend::Sqlite,
-        storage_path: Some(resolve_sqlite_path(root_dir)),
-        reason_code: MEMORY_STORAGE_REASON_DEFAULT_SQLITE.to_string(),
-    }
-}
-
-fn resolve_jsonl_path(root_dir: &Path) -> PathBuf {
-    match root_dir
-        .extension()
-        .and_then(|value| value.to_str())
-        .map(|value| value.to_ascii_lowercase())
-        .as_deref()
-    {
-        Some("jsonl") => root_dir.to_path_buf(),
-        _ => root_dir.join(MEMORY_RUNTIME_ENTRIES_FILE_NAME),
-    }
-}
-
-fn resolve_sqlite_path(root_dir: &Path) -> PathBuf {
-    match root_dir
-        .extension()
-        .and_then(|value| value.to_str())
-        .map(|value| value.to_ascii_lowercase())
-        .as_deref()
-    {
-        Some("sqlite" | "db") => root_dir.to_path_buf(),
-        _ => root_dir.join(MEMORY_RUNTIME_ENTRIES_SQLITE_FILE_NAME),
-    }
-}
-
-fn looks_like_sqlite_file(path: &Path) -> bool {
-    if !path.exists() || !path.is_file() {
-        return false;
-    }
-    let mut file = match fs::File::open(path) {
-        Ok(file) => file,
-        Err(_) => return false,
-    };
-    let mut prefix = [0u8; 16];
-    if std::io::Read::read(&mut file, &mut prefix).unwrap_or_default() < 16 {
-        return false;
-    }
-    &prefix == b"SQLite format 3\0"
-}
-
-fn append_record_jsonl(path: &Path, record: &RuntimeMemoryRecord) -> Result<()> {
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent)
-            .with_context(|| format!("failed to create memory store root {}", parent.display()))?;
-    }
-
-    let mut file = OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(path)
-        .with_context(|| format!("failed to open memory entries file {}", path.display()))?;
-    let encoded = serde_json::to_string(record).context("failed to encode memory record")?;
-    file.write_all(encoded.as_bytes())
-        .with_context(|| format!("failed to write memory record to {}", path.display()))?;
-    file.write_all(b"\n")
-        .with_context(|| format!("failed to write newline to {}", path.display()))?;
-    file.flush()
-        .with_context(|| format!("failed to flush memory entries file {}", path.display()))?;
-    Ok(())
-}
-
-fn load_records_jsonl(path: &Path) -> Result<Vec<RuntimeMemoryRecord>> {
-    if !path.exists() {
-        return Ok(Vec::new());
-    }
-    let file = fs::File::open(path)
-        .with_context(|| format!("failed to open memory entries file {}", path.display()))?;
-    let reader = BufReader::new(file);
-    let mut records = Vec::new();
-    for (index, line) in reader.lines().enumerate() {
-        let line = line.with_context(|| {
-            format!(
-                "failed to read memory entries file {} at line {}",
-                path.display(),
-                index + 1
-            )
-        })?;
-        let trimmed = line.trim();
-        if trimmed.is_empty() {
-            continue;
-        }
-        let record = serde_json::from_str::<RuntimeMemoryRecord>(trimmed).with_context(|| {
-            format!(
-                "failed to parse memory entries file {} at line {}",
-                path.display(),
-                index + 1
-            )
-        })?;
-        records.push(record);
-    }
-    Ok(records)
-}
-
-fn append_record_sqlite(path: &Path, record: &RuntimeMemoryRecord) -> Result<()> {
-    let connection = open_memory_sqlite_connection(path)?;
-    initialize_memory_sqlite_schema(&connection)?;
-    let encoded = serde_json::to_string(record).context("failed to encode memory record")?;
-    connection.execute(
-        r#"
-        INSERT INTO memory_records (memory_id, updated_unix_ms, record_json)
-        VALUES (?1, ?2, ?3)
-        "#,
-        params![record.entry.memory_id, record.updated_unix_ms, encoded],
-    )?;
-    Ok(())
-}
-
-fn load_records_sqlite(path: &Path) -> Result<Vec<RuntimeMemoryRecord>> {
-    if !path.exists() {
-        return Ok(Vec::new());
-    }
-    let connection = open_memory_sqlite_connection(path)?;
-    initialize_memory_sqlite_schema(&connection)?;
-    let mut statement = connection.prepare(
-        r#"
-        SELECT record_json
-        FROM memory_records
-        ORDER BY row_id ASC
-        "#,
-    )?;
-    let mut rows = statement.query([])?;
-    let mut records = Vec::new();
-    while let Some(row) = rows.next()? {
-        let encoded: String = row.get(0)?;
-        let record = serde_json::from_str::<RuntimeMemoryRecord>(&encoded)
-            .context("failed to decode sqlite memory record")?;
-        records.push(record);
-    }
-    Ok(records)
-}
-
-fn open_memory_sqlite_connection(path: &Path) -> Result<Connection> {
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent)
-            .with_context(|| format!("failed to create memory store root {}", parent.display()))?;
-    }
-    let connection = Connection::open(path)
-        .with_context(|| format!("failed to open sqlite memory store {}", path.display()))?;
-    connection.busy_timeout(std::time::Duration::from_secs(5))?;
-    connection.execute_batch(
-        r#"
-        PRAGMA journal_mode = WAL;
-        PRAGMA synchronous = NORMAL;
-        "#,
-    )?;
-    Ok(connection)
-}
-
-fn initialize_memory_sqlite_schema(connection: &Connection) -> Result<()> {
-    connection.execute_batch(
-        r#"
-        CREATE TABLE IF NOT EXISTS memory_records (
-            row_id INTEGER PRIMARY KEY AUTOINCREMENT,
-            memory_id TEXT NOT NULL,
-            updated_unix_ms INTEGER NOT NULL,
-            record_json TEXT NOT NULL
-        );
-        CREATE INDEX IF NOT EXISTS idx_memory_records_memory_id_updated
-            ON memory_records(memory_id, updated_unix_ms);
-        "#,
-    )?;
-    Ok(())
 }
 
 fn current_unix_timestamp_ms() -> u64 {
